requirements.txt
asgiref==3.4.1
astroid==2.8.5
cffi==1.15.0
coverage==6.1.2
cryptography==35.0.0
ddt==1.4.4
Django==3.2.8
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
importlib-metadata==4.8.2
isort==5.10.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
pylint==2.11.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.5.0
typing-extensions==3.10.0.2
wrapt==1.13.3
zipp==3.6.0
